my long-winded fight for running es8 tests has concluded, an es7 and es8 instance are included in our github action matrices, and tests are passing in es7 and es8 (except ones relying on mapping types, which aren't supported by es8)
cleanup (+15 squashed commits)
Squashed commits:
[3fb4418] start excluding mapping types in es8 tests again
[4dc2703] alright separating and cleaning
[34590cc] I want to check something cheeky
[2fec4a1] flip reverse it so only es7 runs mapping type tests
[ce87d3d] getting weird with it
[fc17334] ok next level
[3ce0564] first pass on a matrix
[7a88a3a] bringing back es8
[bb819f6] bringing the port back
[224e686] skipping es8 tests, not running its action
[0e46124] whatever I'll just skip those tests again
[c0afa77] do I get a different error at least
[5842e27] I call this throwing stuff at the wall and hoping for the best
[364c0e1] container name overlap
[eb6d57a] es8 in gh actions
