+++ config/test.exs
@@ -2,10 +2,7 @@ import Config
 
 config :elastix, json_codec: Jason
 
-config :exlasticsearch, ExlasticSearch.Repo,
-  url: "http://localhost:9200",
-  es8: "http://localhost:9201"
-
+config :exlasticsearch, ExlasticSearch.Repo, url: "http://localhost:9200"
 config :exlasticsearch, ExlasticSearch.TypelessTestModel, index_type: :es8
 config :exlasticsearch, :monitoring, ExlasticSearch.Monitoring.Mock
