--- docker-compose.yaml
@@ -1,24 +1,0 @@
-version: "3.7"
-volumes:
-  elasticsearch-7-data:
-  elasticsearch-8-data:
-services:
-  elasticsearch-7:
-    image: docker.elastic.co/elasticsearch/elasticsearch:7.17.16
-    restart: always
-    environment:
-      discovery.type: single-node
-    volumes:
-      - elasticsearch-7-data:/usr/share/elasticsearch/data
-    ports:
-      - 9200:9200
-  elasticsearch-8:
-    image: docker.elastic.co/elasticsearch/elasticsearch:8.13.4
-    restart: always
-    environment:
-      discovery.type: single-node
-    volumes:
-      - elasticsearch-8-data:/usr/share/elasticsearch/data
-      - ./elasticsearch.yml:/usr/share/elasticsearch/config/elasticsearch.yml:ro
-    ports:
-      - 9201:9200
